Операнды double, а не int

+++ CliCalculator/Calculator/Calculator.cs
@@ -12,7 +12,7 @@ namespace CliCalculator.Calculator
 
         private readonly IToken[] tokens;
 
-        public int ResolveExpression()
+        public double ResolveExpression()
         {
             var stack = new Stack<IToken>();
 
@@ -46,7 +46,7 @@ namespace CliCalculator.Calculator
                                 stack.Push(new OperandToken(leftOperand.Number / rightOperand.Number));
                                 break;
                             case OperatorType.Power:
-                                stack.Push(new OperandToken((int)Math.Pow(leftOperand.Number, rightOperand.Number)));
+                                stack.Push(new OperandToken(Math.Pow(leftOperand.Number, rightOperand.Number)));
                                 break;
                         }
                         break;
